requirements.txt
# Capture - Screenshot Enhancement Tool
# Core Dependencies

# GUI Framework
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.1.0

# Database
SQLAlchemy>=2.0.23

# OCR for PII Detection
pytesseract>=0.3.10

# File type validation
python-magic>=0.4.27

# Utilities
python-dateutil>=2.8.2
